refactor: Grid screens use abstract WynntilsGridLayoutScreen
